Change user input valid check to "&&" instead of "||"
When evaluating if the user input does not equal 'Rock' or 'Paper or 'Scissors', the evaluation will always be true, as there is only one prompt, it will always not equal one of the other two.
Using "&&" instead correctly evaluates true and resets the game ONLY if a string other than those three is entered.