Fix broken loop when player input is invalid.
Previous code reran playRound() everytime the player input is invalid. This would cause the rest of the game logic to run with the invalid input, AND ALSO call another round,
instead of the intended behaviour of prompting the player again, and reassigning the variable playerSelection BEFORE continuing with the current round.
New playerSelectionPrompt() function is now run to prompt for the player input, and on invalid input, will call upon itself again, until the variable playerSelection is assigned a valid input, before continuing the round.